pyproject.toml
[project]
name = "lite-rt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "notebook>=7.5.2",
    "numpy<2",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "tabulate>=0.9.0",
    "tflite-runtime>=2.14.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
